requirements.txt
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0

pandas
crayons==0.4.0
PyPDF2
pypdf
python-docx
docx2txt